Removed the Restricted interface and special permission handling for admin role.

+++ src/java/org/rapidcontext/app/proc/PluginInstallProcedure.java
@@ -1,6 +1,6 @@
 /*
  * RapidContext <http://www.rapidcontext.com/>
- * Copyright (c) 2007-2012 Per Cederberg. All rights reserved.
+ * Copyright (c) 2007-2013 Per Cederberg. All rights reserved.
  *
  * This program is free software: you can redistribute it and/or
  * modify it under the terms of the BSD license.
@@ -22,8 +22,6 @@ import org.rapidcontext.core.proc.Bindings;
 import org.rapidcontext.core.proc.CallContext;
 import org.rapidcontext.core.proc.Procedure;
 import org.rapidcontext.core.proc.ProcedureException;
-import org.rapidcontext.core.security.Restricted;
-import org.rapidcontext.core.security.SecurityContext;
 import org.rapidcontext.core.type.Session;
 
 /**
@@ -32,7 +30,7 @@ import org.rapidcontext.core.type.Session;
  * @author   Per Cederberg
  * @version  1.0
  */
-public class PluginInstallProcedure implements Procedure, Restricted {
+public class PluginInstallProcedure implements Procedure {
 
     /**
      * The procedure name constant.
@@ -53,15 +51,6 @@ public class PluginInstallProcedure implements Procedure, Restricted {
         defaults.set("sessionFileId", Bindings.ARGUMENT, "",
                      "The session file identifier (containing the plug-in zip file).");
         defaults.seal();
-    }
-
-    /**
-     * Checks if the currently authenticated user has access to this object.
-     *
-     * @return true if the current user has access, or false otherwise
-     */
-    public boolean hasAccess() {
-        return SecurityContext.hasAdmin();
     }
 
     /**
@@ -126,6 +115,7 @@ public class PluginInstallProcedure implements Procedure, Restricted {
             msg = "failed to read session file with id '" + fileId + "'";
             throw new ProcedureException(msg);
         }
+        CallContext.checkWriteAccess("plugin/" + fileId);
         try {
             pluginId = ctx.installPlugin(file);
             session.removeFile(fileId);
