+++ src/java/org/rapidcontext/app/proc/ProcedureDeleteProcedure.java
@@ -1,6 +1,6 @@
 /*
  * RapidContext <http://www.rapidcontext.com/>
- * Copyright (c) 2007-2012 Per Cederberg. All rights reserved.
+ * Copyright (c) 2007-2013 Per Cederberg. All rights reserved.
  *
  * This program is free software: you can redistribute it and/or
  * modify it under the terms of the BSD license.
@@ -18,8 +18,6 @@ import org.rapidcontext.core.proc.Bindings;
 import org.rapidcontext.core.proc.CallContext;
 import org.rapidcontext.core.proc.Procedure;
 import org.rapidcontext.core.proc.ProcedureException;
-import org.rapidcontext.core.security.Restricted;
-import org.rapidcontext.core.security.SecurityContext;
 
 /**
  * The built-in procedure delete procedure.
@@ -27,7 +25,7 @@ import org.rapidcontext.core.security.SecurityContext;
  * @author   Per Cederberg
  * @version  1.0
  */
-public class ProcedureDeleteProcedure implements Procedure, Restricted {
+public class ProcedureDeleteProcedure implements Procedure {
 
     /**
      * The procedure name constant.
@@ -47,17 +45,6 @@ public class ProcedureDeleteProcedure implements Procedure, Restricted {
     public ProcedureDeleteProcedure() throws ProcedureException {
         defaults.set("name", Bindings.ARGUMENT, "", "The procedure name");
         defaults.seal();
-    }
-
-    /**
-     * Checks if the currently authenticated user has access to this
-     * object.
-     *
-     * @return true if the current user has access, or
-     *         false otherwise
-     */
-    public boolean hasAccess() {
-        return SecurityContext.hasAdmin();
     }
 
     /**
@@ -114,6 +101,7 @@ public class ProcedureDeleteProcedure implements Procedure, Restricted {
         if (name.length() == 0) {
             throw new ProcedureException("invalid procedure name");
         }
+        CallContext.checkWriteAccess("procedure/" + name);
         cx.getLibrary().deleteProcedure(name);
         return null;
     }
